neon: Refactor the macros for implementing rules
Basically, convert the "BINARY()", "UNARY()", ... macros into functions
that receives a "OrcNeonInsn" structure with the required data. Instead of
having a lot of identical functions with a different data in the stack,
move all these data into a static table and reuse the same function.

The check "if (insn_name64)" was deleted, as currently there is no case
where this is true.

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/259>